oracle-jse-jre: fix QA warning on shlib provider
Fix this QA warning:
WARNING: glibc-2.22 was registered as shlib provider for ld-linux.so.3,
changing it to oracle-jse-jre-1.7.0 because it was built later

Java is not provider of ld-linux.so.3, it is glibc. Mark it as a private
lib to fix this.

Also fix comment (symlink was done in previous version).

Signed-off-by: Tomas Novotny <tomas@novotny.cz>